refactor fx solution into other folders

+++ dotnet/fx/Standard/test/Std/Fs_Tests.cs
@@ -4,43 +4,50 @@ namespace Test.Std;
 
 public class Fs_Tests
 {
-    private static readonly string Csproj = FsPath.Combine(Util.StandardDir, "src", "Bearz.Standard.csproj");
-
     [IntegrationTest]
     public void Verify_Attr(IAssert assert)
     {
-        var attrs = Fs.Attr(Util.StandardDir);
+        var bearzDir = CreateTestDir();
+        var textFile = FsPath.Combine(bearzDir, "alpha", "test.txt");
+        var attrs = Fs.Attr(bearzDir);
         assert.True(attrs.HasFlag(FileAttributes.Directory));
 
-        attrs = Fs.Attr(Csproj);
+        attrs = Fs.Attr(textFile);
         assert.False(attrs.HasFlag(FileAttributes.Directory));
     }
 
     [IntegrationTest]
     public void Verify_Stat(IAssert assert)
     {
-        var fsi = Fs.Stat(Util.StandardDir);
+        var bearzDir = CreateTestDir();
+        var fsi = Fs.Stat(bearzDir);
         assert.True(fsi is DirectoryInfo);
     }
 
     [IntegrationTest]
     public void Verify_IsDirectory(IAssert assert)
     {
-        assert.True(Fs.IsDirectory(Util.StandardDir));
-        assert.False(Fs.IsDirectory(Csproj));
+        var bearzDir = CreateTestDir();
+        var textFile = FsPath.Combine(bearzDir, "alpha", "test.txt");
+        assert.True(Fs.IsDirectory(bearzDir));
+        assert.False(Fs.IsDirectory(textFile));
     }
 
     [IntegrationTest]
     public void Verify_IsFile(IAssert assert)
     {
-        assert.False(Fs.IsFile(Util.StandardDir));
-        assert.True(Fs.IsFile(Csproj));
+        var bearzDir = CreateTestDir();
+        var textFile = FsPath.Combine(bearzDir, "alpha", "test.txt");
+        assert.False(Fs.IsFile(bearzDir));
+        assert.True(Fs.IsFile(textFile));
     }
 
     [IntegrationTest]
     public void Verify_Open(IAssert assert)
     {
-        using var stream = Fs.Open(Csproj);
+        var bearzDir = CreateTestDir();
+        var textFile = FsPath.Combine(bearzDir, "alpha", "test.txt");
+        using var stream = Fs.Open(textFile);
         assert.True(stream.CanRead);
         assert.True(stream.Length > 0);
     }
@@ -48,10 +55,12 @@ public class Fs_Tests
     [IntegrationTest]
     public void Verify_MakeAndDeleteDirectory(IAssert assert)
     {
-        var tmp = FsPath.TempDir();
+        var tmp = FsPath.GetTempDir();
         var dir = FsPath.Combine(tmp, "foo", "bar");
         var parent = FsPath.Dirname(dir)!;
-        Fs.RemoveDirectory(parent, true);
+        if (Fs.DirectoryExists(parent))
+            Fs.RemoveDirectory(parent, true);
+
         assert.False(Fs.DirectoryExists(dir));
         Fs.MakeDirectory(dir);
         assert.True(Fs.DirectoryExists(dir));
@@ -63,8 +72,8 @@ public class Fs_Tests
     [IntegrationTest]
     public void Verify_CopyDirectory(IAssert assert)
     {
-        var tmp = FsPath.TempDir();
-        var src = FsPath.Combine(Util.StandardDir);
+        var tmp = FsPath.GetTempDir();
+        var bearzDir = CreateTestDir();
         var dst = FsPath.Combine(tmp, "dst");
         if (Fs.DirectoryExists(dst))
             Fs.RemoveDirectory(dst, true);
@@ -72,19 +81,36 @@ public class Fs_Tests
         try
         {
             assert.False(Fs.DirectoryExists(dst));
-            Fs.CopyDirectory(src, dst, true, false);
+            Fs.CopyDirectory(bearzDir, dst, true);
 
             assert.True(Fs.DirectoryExists(dst));
-            assert.True(Fs.DirectoryExists(FsPath.Combine(dst)));
-            assert.True(Fs.DirectoryExists(FsPath.Combine(dst, "src")));
-            assert.True(Fs.DirectoryExists(FsPath.Combine(dst, "test")));
-            assert.True(Fs.DirectoryExists(FsPath.Combine(dst, "test", "Std")));
-            assert.True(Fs.FileExists(FsPath.Combine(dst, "test", "Std", "Fs_Tests.cs")));
+            assert.True(Fs.DirectoryExists(FsPath.Combine(dst, "alpha")));
+            assert.True(Fs.FileExists(FsPath.Combine(dst, "alpha", "test.txt")));
+            assert.True(Fs.DirectoryExists(FsPath.Combine(dst, "foo")));
+            assert.True(Fs.DirectoryExists(FsPath.Combine(dst, "foo", "bar")));
         }
         finally
         {
             if (Fs.DirectoryExists(dst))
                 Fs.RemoveDirectory(dst, true);
         }
+    }
+
+    private static string CreateTestDir()
+    {
+        var tmp = Path.GetTempPath();
+        var dir = Path.Combine(tmp, "bearz", "foo", "bar");
+        if (!Directory.Exists(dir))
+            Directory.CreateDirectory(dir);
+
+        dir = Path.Combine(tmp, "bearz", "alpha");
+        if (!Directory.Exists(dir))
+            Directory.CreateDirectory(dir);
+
+        var testFile = Path.Combine(dir, "test.txt");
+        if (!File.Exists(testFile))
+            File.WriteAllText(testFile, "test");
+
+        return Path.Combine(tmp, "bearz");
     }
 }
